add images and principal grill

+++ MVCBlog.Core/Entities/PrecioHistorico.cs
@@ -8,7 +8,7 @@ namespace MVCBlog.Core.Entities
 {
     public class PrecioHistorico : EntityBase
     {
-        public decimal Precio { get; set; }
+        public int Precio { get; set; }
 
         public DateTime FechaDesde { get; set; }
